% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figS1panelB.r
\name{figS1panelB}
\alias{figS1panelB}
\title{Generate graph of figure S1 lower panel}
\usage{
figS1panelB()
}
\description{
Generate graph of figure S1 lower panel
}
\examples{
figS1panelB()

}
\author{
JuG
}
