% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDecay.r
\name{fitDecay}
\alias{fitDecay}
\title{Do something}
\usage{
fitDecay(matDecayArray, Na = 50, start = 100, A = 1000,
  nchannels = 1024)
}
\arguments{
\item{matDecayArray}{matrix of the decays}

\item{Na}{= 50,}

\item{start}{channel at which the decay starts}

\item{A}{amplitude of the decay}

\item{nchannels}{Number of channels (default = 1024)}
}
\description{
Do something
}
\examples{

fitDecay(matDecayArray = decay)
}
\author{
JuG
}
