\name{flimDiagRam-package}
\alias{flimDiagRam-package}
\alias{flimDiagRam}
\docType{package}
\title{
\packageTitle{flimDiagRam}
}
\description{
\packageDescription{flimDiagRam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{flimDiagRam}
\packageIndices{flimDiagRam}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{flimDiagRam}

Maintainer: \packageMaintainer{flimDiagRam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
