% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.r
\name{getData}
\alias{getData}
\title{Read single exponential decay data file}
\usage{
getData(pathTau, pathPhoton, label, minPhotons = 0)
}
\arguments{
\item{pathTau}{path to the lifetime data file (*_t1.asc)}

\item{pathTau}{path to the lifetime data file (*_photons.asc)}
}
\value{
data frame
}
\description{
Read single exponential decay data file
}
\examples{


}
\author{
JuG
}
