% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataHist.R
\name{getDataHist}
\alias{getDataHist}
\title{Import data}
\usage{
getDataHist(pathEGFP, label)
}
\arguments{
\item{pathEGFP}{path to folder containing _a1[\%].asc, _photons.asc and _t1.asc files}

\item{label}{labdel string}
}
\description{
Import data
}
\examples{


}
\author{
JuG
}
