% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvkde.r
\name{mvkde}
\alias{mvkde}
\title{Compute KDE at x given data}
\usage{
mvkde(x, data, bandwidth = Scott, kernel = RadSym)
}
\description{
Published by  <matt.shotwell@vanderbilt.edu> https://biostatmatt.com/archives/2745
}
\details{

}
\author{
matt.shotwell
}
