% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecay.r
\name{plotDecay}
\alias{plotDecay}
\title{Graphical plot of an individual decay}
\usage{
plotDecay(decay, channels, timePerChannel = 0.0125, fit,
  addPoints = FALSE, ...)
}
\arguments{
\item{decay}{photon counts of the decay}

\item{channels}{channels vector}

\item{timePerChannel}{time per channels (by default 0.0125ns = 12.5ps)}

\item{addPoints}{boolean add decay on plotted graph}
}
\description{
Graphical plot of an individual decay
}
\examples{
decay <- simMatDecay(nAcceptor = 1,nAcceptorFix=TRUE, fraction = "progressive", 
distAcc = 35.383045, sdDistAcc = 0,bindingProbs = 1)
plotDecay(decay = decay[1,80,],col='blue')
plotDecay(decay = decay[1,10,], add=TRUE)

}
\author{
JuG
}
