% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEcdf.r
\name{plotEcdf}
\alias{plotEcdf}
\title{Plot empirical cumulative distribution functions}
\usage{
plotEcdf(AA, BB, col1, col2, ...)
}
\arguments{
\item{AA}{dataset a}

\item{BB}{dataset b}
}
\description{
Plot empirical cumulative distribution functions
}
\details{

}
\examples{


}
\author{
JuG
}
