% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadSym.r
\name{RadSym}
\alias{RadSym}
\title{Radially symmetric kernel (Gaussian kernel)}
\usage{
RadSym(u)
}
\description{
Published by  <matt.shotwell@vanderbilt.edu> https://biostatmatt.com/archives/2745
}
\examples{

}
\author{
<jgodet@unistra.fr> (adapted from matt.shotwell)
}
