% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scott.r
\name{Scott}
\alias{Scott}
\title{Multivariate extension of Scott's bandwidth rule}
\usage{
Scott(data)
}
\description{
Published by  <matt.shotwell@vanderbilt.edu> https://biostatmatt.com/archives/2745
}
\details{

}
\author{
matt.shotwell
}
