% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smvkde.r
\name{smvkde}
\alias{smvkde}
\title{Compute KDE at (matrix) x given data}
\usage{
smvkde(x, ...)
}
\description{
Published by  <matt.shotwell@vanderbilt.edu> https://biostatmatt.com/archives/2745
}
\examples{


}
\author{
matt.shotwell
}
