% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagram.r
\name{plotDiagram}
\alias{plotDiagram}
\title{Plot FLIM diagram}
\usage{
plotDiagram(data, ...)
}
\arguments{
\item{data}{dataframe of alpha and tau values}
}
\description{
Plot FLIM diagram
}
\examples{


}
\author{
JuG
}
